requirements.txt
# Core Dependencies
click>=8.0.0
pydantic>=2.4.0
PyYAML>=6.0.1
requests>=2.31.0
jinja2>=3.1.5

# Cloud Provider SDKs
boto3>=1.35.0
azure-mgmt-core>=1.4.0
azure-mgmt-resource>=23.1.0
azure-mgmt-storage>=21.1.0
azure-mgmt-compute>=30.1.0
google-cloud-storage>=2.12.0
google-cloud-resource-manager>=1.12.0
google-cloud-securitycenter>=1.26.0

# Security and Vulnerability Scanning
checkov>=3.2.50
bandit>=1.7.6
safety>=3.0.1
semgrep>=1.50.0

# Container and Kubernetes Security
docker>=7.0.0
kubernetes>=29.0.0

# SBOM Generation and Analysis
cyclonedx-bom>=4.1.0
packageurl-python>=0.11.2

# AI/ML for Vulnerability Prioritization
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.25.2

# Database and Caching
sqlalchemy>=2.0.25
redis>=5.0.1

# Web Framework for Dashboard
fastapi>=0.108.0
uvicorn>=0.25.0

# Reporting and Visualization
reportlab>=4.0.8
matplotlib>=3.8.2
plotly>=5.17.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0 